+++ src/templates/globals.ts
@@ -1,0 +1,3 @@
+import Template from "./template";
+
+export const ObjectDefineProperty = Template(`Object["defineProperty"]`);