+++ src/transforms/identifier/globalConcealing.ts
@@ -25,6 +25,7 @@ import { chance, getRandomInteger } from "../../util/random";
 import { reservedIdentifiers } from "../../constants";
 import { ComputeProbabilityMap } from "../../probability";
 import GlobalAnalysis from "./globalAnalysis";
+import { GetGlobalTemplate } from "../../templates/bufferToString";
 
 /**
  * Global Concealing hides global variables being accessed.
@@ -72,42 +73,19 @@ export default class GlobalConcealing extends Transform {
       });
 
       if (Object.keys(globals).length > 0) {
-        var used = new Set();
+        var usedStates = new Set<number>();
 
         // Make getter function
 
         // holds "window" or "global"
         var globalVar = this.getPlaceholder();
 
-        // holds outermost "this"
-        var thisVar = this.getPlaceholder();
-
-        // "window" or "global" in node
-        var global =
-          this.options.globalVariables.values().next().value || "window";
-        var alternateGlobal = global === "window" ? "global" : "window";
-
         var getGlobalVariableFnName = this.getPlaceholder();
-        var getThisVariableFnName = this.getPlaceholder();
 
         // Returns global variable or fall backs to `this`
-        var getGlobalVariableFn = Template(`
-        var ${getGlobalVariableFnName} = function(){
-          try {
-            return ${global} || ${alternateGlobal} || (new Function("return this"))();
-          } catch (e){
-            return ${getThisVariableFnName}["call"](this);
-          }
-        }`).single();
-
-        var getThisVariableFn = Template(`
-        var ${getThisVariableFnName} = function(){
-          try {
-            return this;
-          } catch (e){
-            return null;
-          }
-        }`).single();
+        var getGlobalVariableFn = GetGlobalTemplate.compile({
+          getGlobalFnName: getGlobalVariableFnName,
+        });
 
         // 2. Replace old accessors
         var globalFn = this.getPlaceholder();
@@ -116,11 +94,11 @@ export default class GlobalConcealing extends Transform {
 
         Object.keys(globals).forEach((name) => {
           var locations: Location[] = globals[name];
-          var state;
+          var state: number;
           do {
-            state = getRandomInteger(-1000, 1000 + used.size);
-          } while (used.has(state));
-          used.add(state);
+            state = getRandomInteger(-1000, 1000 + usedStates.size);
+          } while (usedStates.has(state));
+          usedStates.add(state);
 
           newNames[name] = state;
 
@@ -139,10 +117,10 @@ export default class GlobalConcealing extends Transform {
             do {
               state = getRandomInteger(
                 0,
-                1000 + used.size + this.options.globalVariables.size * 100
+                1000 + usedStates.size + this.options.globalVariables.size * 100
               );
-            } while (used.has(state));
-            used.add(state);
+            } while (usedStates.has(state));
+            usedStates.add(state);
 
             newNames[name] = state;
           }
@@ -162,15 +140,7 @@ export default class GlobalConcealing extends Transform {
                 var code = newNames[name];
                 var body: Node[] = [
                   ReturnStatement(
-                    LogicalExpression(
-                      "||",
-                      MemberExpression(
-                        Identifier(globalVar),
-                        Literal(name),
-                        true
-                      ),
-                      MemberExpression(Identifier(thisVar), Literal(name), true)
-                    )
+                    MemberExpression(Identifier(globalVar), Literal(name), true)
                   ),
                 ];
                 if (chance(50)) {
@@ -183,7 +153,7 @@ export default class GlobalConcealing extends Transform {
                           "||",
                           Literal(name),
                           MemberExpression(
-                            Identifier(thisVar),
+                            Identifier(globalVar),
                             Literal(name),
                             true
                           )
@@ -198,18 +168,10 @@ export default class GlobalConcealing extends Transform {
               })
             ),
             ReturnStatement(
-              LogicalExpression(
-                "||",
-                MemberExpression(
-                  Identifier(globalVar),
-                  Identifier(returnName),
-                  true
-                ),
-                MemberExpression(
-                  Identifier(thisVar),
-                  Identifier(returnName),
-                  true
-                )
+              MemberExpression(
+                Identifier(globalVar),
+                Identifier(returnName),
+                true
               )
             ),
           ]
@@ -218,7 +180,7 @@ export default class GlobalConcealing extends Transform {
         var tempVar = this.getPlaceholder();
 
         var variableDeclaration = Template(`
-        var ${globalVar}, ${thisVar};
+        var ${globalVar};
         `).single();
 
         variableDeclaration.declarations.push(
@@ -229,11 +191,9 @@ export default class GlobalConcealing extends Transform {
                 FunctionExpression(
                   [],
                   [
-                    getGlobalVariableFn,
-                    getThisVariableFn,
-
+                    ...getGlobalVariableFn,
                     Template(
-                      `return ${thisVar} = ${getThisVariableFnName}["call"](this, ${globalFn}), ${globalVar} = ${getGlobalVariableFnName}["call"](this)`
+                      `return ${globalVar} = ${getGlobalVariableFnName}["call"](this)`
                     ).single(),
                   ]
                 ),